+++ client/src/components/Visualization/charts/Geochart.js
@@ -1,10 +1,21 @@
-import { ads, states } from '../../../firebase/GetFirestoreDocuments';
+import { ads } from '../../../firebase/FirestoreDocumentReader';
 import { Chart } from "react-google-charts";
 import firebase from '../../../firebase/firebase';
 import React from 'react';
+import {states } from './StateDataParser';
+
+/* 
+ * Description: Geochart implements a Google geochart component.
+ * Author: Kira Toal
+ * Date: 2020/07/13
+ */
 
 const Geochart = () => {
-
+  /*
+   * Currently, getData assigns a random number to each state.
+   * In the future, once the states dictionary contains ad information, getData will
+   *    retrieve information such as minimum ad impressions and maximum ad spend.
+   */ 
   function getData() {
     let data = [["State", "Random Number"]];
     for (let state in states) {
